refactor-#156: move customer controller to controllers
