+++ src/microservices/customer/customer.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { CustomerController } from './rest/customer.controller';
 import { ClientsModule } from '@nestjs/microservices';
 
 import { createClientAsyncOptions } from '@utils/client';
+import { CustomerController } from './rest/controllers';
 import { CustomerService } from './services';
 import { CustomerQueryResolver } from './graphql/customer-query.resolver';
 import { CustomerMutationResolver } from './graphql/customer-mutation.resolver';
