+++ src/microservices/customer/rest/controllers/customer.controller.spec.ts
@@ -2,7 +2,7 @@ import { ClientsModule } from '@nestjs/microservices';
 import { Test, TestingModule } from '@nestjs/testing';
 
 import { createClientAsyncOptions } from '@utils/client';
-import { CustomerService } from '../services';
+import { CustomerService } from '../../services';
 import { CustomerController } from './customer.controller';
 
 describe('CustomerController', () => {
